Separate hydra utils Fixes #46
